+++ packages/components/src/components/atoms/form/DropDown.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { renderWithProviders } from '../../../utils/tests'
 
-import DropDownError from './DropDownError'
+import DropDown from './DropDown'
 import { screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
@@ -24,38 +24,38 @@ const options = [
 describe('DropDownError', () => {
   it('sould match snapshot', () => {
     const { container } = renderWithProviders(
-      <DropDownError options={options} value={1} />
+      <DropDown options={options} value={1} />
     )
     expect(container).toMatchSnapshot()
   })
 
   it('sould match snapshot width multiple value', () => {
     const { container } = renderWithProviders(
-      <DropDownError options={options} multiple value={[1, 2]} />
+      <DropDown options={options} multiple value={[1, 2]} />
     )
     expect(container).toMatchSnapshot()
   })
 
   it('should display error with showError prop when the field is required with missing value', () => {
-    renderWithProviders(<DropDownError options={options} required showError />)
+    renderWithProviders(<DropDown options={options} required showError />)
     expect(screen.getByText('formError.valueMissing')).toBeInTheDocument()
   })
 
   it('should not display error without showError prop when the field is required with missing value', () => {
-    renderWithProviders(<DropDownError options={options} required />)
+    renderWithProviders(<DropDown options={options} required />)
     expect(screen.queryByText('formError.valueMissing')).not.toBeInTheDocument()
   })
 
   it('should not display error with showError prop when the field is required and disabled with missing value', () => {
     renderWithProviders(
-      <DropDownError options={options} required showError disabled />
+      <DropDown options={options} required showError disabled />
     )
     expect(screen.queryByText('formError.valueMissing')).not.toBeInTheDocument()
   })
 
   it('shoud display error with showError prop when the field has additional validation rules', () => {
     renderWithProviders(
-      <DropDownError
+      <DropDown
         options={options}
         required
         showError
@@ -67,7 +67,7 @@ describe('DropDownError', () => {
 
   it('should display a replacement error message with replacementErrorsMessages prop when the field has an error', () => {
     renderWithProviders(
-      <DropDownError
+      <DropDown
         options={options}
         required
         showError
