+++ packages/components/src/components/index.ts
@@ -10,8 +10,8 @@ export { default as DatePickerWithoutError } from './atoms/form/DatePickerWithou
 export { default as DatePicker } from './atoms/form/DatePicker'
 export { default as DoubleDatePickerWithoutError } from './atoms/form/DoubleDatePickerWithoutError'
 export { default as DoubleDatePicker } from './atoms/form/DoubleDatePicker'
+export { default as DropDownWithoutError } from './atoms/form/DropDownWithoutError'
 export { default as DropDown } from './atoms/form/DropDown'
-export { default as DropDownError } from './atoms/form/DropDownError'
 export { default as InfoTooltip } from './atoms/form/InfoTooltip'
 export { default as InputText } from './atoms/form/InputText'
 export { default as InputTextError } from './atoms/form/InputTextError'
